MTP-1970: Added before argument to `delete_old_data` command
This is optional and defaults to 7 years ago as before. There is also
a failsafe that prevents accidentally passing a date newer than 7 years
ago (e.g. to prevent deletion of newer data).

This could be useful to stagger the deletion of old data given the first
run would be slow with a lot of very old credits.
Eventually we could run this daily and the amount of data deleted should
be much more managable.