Upgraded drf-nested-routers from 0.93.5 to 0.94.1
The project don't seem to keep a changelog and the
releases on GH/pypi don't have much info in them which
is not great.

However we weren't using the newer version because
[0.94.1 did drop] [support for Django 3.2] we were using.

I don't seem to be getting any errors and the tests
I've run seem to pass: Module is imported in `credit`,
mtp_auth and security, so hopefully run the views tests
of these apps should highlight any potentially compatibility
issues.

[0.94.1 did drop]: https://github.com/alanjds/drf-nested-routers/commit/b3ecc1211ae5c0a2e6f6ee79a27583250a57cc7d
[support for Django 3.2]: https://github.com/alanjds/drf-nested-routers/commit/f946fd374d120de6838d102fe19d8cb9badc59b3