The `BaseUserManager.make_random_password()` method is removed
`make_random_password()` is removed in Django 5.1.
[Internally] it was just using `django.utils.crypto.get_random_string`
so I've replaced it with the same function call.

NOTE: The default allowed characters in `get_random_string()` are **almost**
the same as the default allowed characters in the old `make_random_password()`
method with the exception of the letter `i`/`I` not being allowed.

I've retained the same behaviour for 2 reason:
1. it shouldn't matter in practice
2. I've made the list of allowed characters explicit in case
   `django.utils.crypto.get_random_string()`'s default change in the future
   for whatever reason.

See: https://docs.djangoproject.com/en/5.1/releases/5.1/#features-removed-in-5-1

[Internally]: https://github.com/django/django/blob/35c58a7924aa07e87e925f6eea978c278f143f36/django/contrib/auth/base_user.py#L36-L51