Remove usage of deprecated `LogEntryManager.log_action()`
As of [Django 5.1] this is deprecated ([will go away in Django 6.0]):

> The `ModelAdmin.log_deletion()` and `LogEntryManager.log_action()`
> methods are deprecated. Subclasses should implement
> `ModelAdmin.log_deletions()` and `LogEntryManager.log_actions()`
> instead.

I've replaced this use with a call to `LogEntry.objects.create()`
which has the same net effect (create a record in the `django_admin_log`
DB table), that's what the deprecated `log_action()` method is
doing internally anyway.

`object_id` and `content_type_id` columns are nullable and we often don't
provide these. While the `change_message` column is not nullable but the
`log_action()` method argument had it default to `''` so I've done the same.

[Django 5.1]: https://docs.djangoproject.com/en/5.2/releases/5.1/#features-deprecated-in-5-1
[will go away in Django 6.0]: https://docs.djangoproject.com/en/5.2/internals/deprecation/#deprecation-removed-in-6-0